+++ frontend/app/api/auth/verify-email/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServiceRoleSupabase } from "@/lib/supabase/server";
-import { isTokenExpired } from "@/lib/utils/auth";
+import { isTokenExpired } from "@/lib/utils/auth-server";
 
 export async function POST(request: NextRequest) {
 	try {
